langChain/requirements.txt
Flask
python-dotenv
pypdf
langchain
langchain-google-genai
langchain-chroma
langchainhub